usbd: kill warnings
cmd_usbd.c: In function 'write_file_mmc':
cmd_usbd.c:376: warning: format '%08x' expects type 'unsigned int', but argument 2 has type 'ulong'
cmd_usbd.c:403: warning: format '%d' expects type 'int', but argument 3 has type 'long unsigned int'
cmd_usbd.c:427: warning: format '%d' expects type 'int', but argument 3 has type 'long unsigned int'

+++ common/cmd_usbd.c
@@ -373,7 +373,7 @@ static void write_file_mmc(struct usbd_ops *usbd, char *ramaddr, u32 len,
 		memcpy(&mbr_info, mbr, sizeof(struct mbr_table));
 
 		printf("Total Size: 0x%08x #parts %d\n",
-				usbd->mmc_total, mmc_parts);
+				(unsigned int)usbd->mmc_total, mmc_parts);
 		for (i = 0; i < mmc_parts; i++) {
 			printf("p%d\t0x%08x\t0x%08x\n", i + 1,
 				mbr_info.partition[i].lba_begin,
@@ -400,7 +400,7 @@ static void write_file_mmc(struct usbd_ops *usbd, char *ramaddr, u32 len,
 		printf("\nWrite Partition %d.. %d blocks\n",
 			cur_partition + 1,
 			part_info.partition[cur_partition].size /
-			usbd->mmc_blk);
+			(int)usbd->mmc_blk);
 	}
 
 	for (i = cur_partition; i < mmc_parts; i++) {
@@ -424,7 +424,7 @@ static void write_file_mmc(struct usbd_ops *usbd, char *ramaddr, u32 len,
 			printf("\nWrite Partition %d.. %d blocks\n",
 				cur_partition + 1,
 				part_info.partition[cur_partition].size /
-				usbd->mmc_blk);
+				(int)usbd->mmc_blk);
 		}
 	}
 }
